package/evemu: needs autoreconf for python 3.12 compatibility
evemu ships with an ancient py-compile script (from 2011) depending
the imp module, which was finally removed in Python 3.12.
Set autoreconf in order to get an up to date version of py-compile
which automatically handles the different python versions.

This is a similar problem as in
42eae73ef1 package/cracklib: python needs autoreconf

Fixes:

 http://autobuild.buildroot.org/results/a52/a520583be6e86993d268d6535f727a352bd92028/

and many other configs where Python3 is selected and thus evemu
is configured to build the python-bindings.

Signed-off-by: Andreas Naumann <anaumann@ultratronik.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>