package/qt6/qt6opcua: new package
Add qt6 version of qtopcua from git as it's not contained in the
official qt release submodule directory.

Fulfill the mandatory dependency on qt6base, and qt6base-network in
particular, as well as the need for host tool qopcuaxmldatatypes2cpp
which is provided by host-qt6opcua.
For host-qt6opcua set the dependency on host-qt6base-network as well.

Existence of the optional dependency openssl enables
FEATURE_open62541_security. (This has changed from qt5opcua where
mbedtls was supported.)

Signed-off-by: Andreas Naumann <anaumann@ultratronik.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>