package/tipidee: fix install without config file
When the tipidee package was added in commit
17deaf60f66fc08a796233920a104e5233b501e0, the committer refactored it so
no separate boolean option is required to enable the config file.
However, in doing so, the case where the config file is empty was not
correctly handled: BR2_PACKAGE_TIPIDEE_CONFIG_FILE is never empty
because even with the empty string it contains "".

Fox by qstrip'ing before checking for empty.

Signed-off-by: Arnout Vandecappelle <arnout@mind.be>