configs/beaglebone_qt5: move patches to beaglebone-qt5 folder
Compared to the beaglebone_defconfig, updating the Linux kernel for
beaglebone_qt5 is more tricky because it may also involve updating the
ti-sgx-{km,um} packages. Therefore, it is unlikely that the two
configurations can share the same software versions for the Linux
kernel, and consequently a single path for the patches. For this
reason, and with the foresight of adding the hash files for the
beaglebone_defconfig as well, the hash files and patches are moved to
a separate board/beagleboard/beaglebone-qt5/patches/ directory. This
way, future conflicts in managing the patches for the two
configurations (i.e., beaglebone[_qt5]_defconfig) are avoided.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>