package/xz: fix musl static build
Runnning autoreconf is needed to avoid the following static build
failure on musl which is raised as otherwise LDFLAGS (which contains
-static) is not passed:

/home/autobuild/autobuild/instance-16/output-1/host/lib/gcc/armeb-buildroot-linux-musleabi/13.3.0/../../../../armeb-buildroot-linux-musleabi/bin/ld: /home/autobuild/autobuild/instance-16/output-1/host/lib/gcc/armeb-buildroot-linux-musleabi/13.3.0/libgcc.a(_dvmd_lnx.o): in function `__aeabi_idiv0':
/home/autobuild/autobuild/instance-16/output-1/build/host-gcc-final-13.3.0/build/armeb-buildroot-linux-musleabi/libgcc/../../../libgcc/config/arm/lib1funcs.S:1499:(.text+0x8): undefined reference to `raise'

This build failure is raised by autobuilders since bump to version 5.6.2
in commit d1d77eb274761d1f2d24ee4b55a374e917c67a8e but it can also be
reproduced on version 5.4.7 (but not on 5.4.6). There is probably an
issue with these two official autotools tarballs which have been
released the same day (on May 29 2024).

Fixes: 40240ac30afc4c6af765794f38e6815162ca61fa

  http://autobuild.buildroot.org/results/344813b202644a23c166920aa7e861ebf408536b

Signed-off-by: Fabrice Fontaine <fontaine.fabrice@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>