package/rsyslog: extra plugins option introduced
The rsyslog package enables extra plugins based on the availability of
external dependencies. However, there are many cases where those
plugins are not needed, even when some of those external dependencies
are enabled. Allow disabling the build of plugins. This option is
useful in embedded systems to achieve compact footprint.

Signed-off-by: Konstantin Menyaev <KAMenyaev@salutedevices.com>
Signed-off-by: Flávio Tapajós <flavio.tapajos@newtesc.com.br>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>