configs/freescale_imx7dsabresd: bump BSP components to lf-6.6.23-2.0.0
This commit bumps U-Boot and Linux kernel to the NXP BSP
lf-6.6.23-2.0.0 versions.

While bumping those BSP components, this commit also adds the custom
hashes by enabling BR2_DOWNLOAD_FORCE_CHECK_HASHES=y in this
defconfig. A _PATCH_DIR is added for that purpose and hash files are
added in it. The ".checkpackageignore" entry for this defconfig is
also removed.

Also, since commit dc0f721 "package/gcc: switch to 13.x as default",
this freescale_imx7dsabresd_defconfig is failing to build the Kernel
with output:

    drivers/ata/libahci.c: In function 'ahci_led_store':
    ././include/linux/compiler_types.h:315:45: error: call to '__compiletime_assert_340' declared with attribute error: BUILD_BUG_ON failed: sizeof(_s) > sizeof(long)
      315 |         _compiletime_assert(condition, msg, __compiletime_assert_, __COUNTER__)
          |                                             ^

This commit fixes the issue by bumping components which includes the
needed gcc-13 fixes.

The Kernel DTS name is also updated by adding the new "nxp/imx/"
prefix, to follow the upstream reorganization in [1], first included
in v6.5.

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/7423016459

[1] https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=724ba6751532055db75992fc6ae21c3e322e94a7

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>