configs/microchip_mpfs_icicle: bump Linux and U-Boot, add hashes
Following Microchip's Linux 2024.06 release, update the assets to the
latest versions, this includes a Linux and U-boot version bump. Update
the Linux headers to support the updated v6.6 kernel.
This defconfig was added before the "BR2_DOWNLOAD_FORCE_CHECK_HASHES"
option was mandatory. Add this and the corresponding package hashes.

Signed-off-by: Jamie Gibbons <jamie.gibbons@microchip.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>