package/imx-mkimage: bump to version lf-6.6.23-2.0.0
For change log, see:
https://github.com/nxp-imx/imx-mkimage/compare/lf-6.1.22-2.0.0...lf-6.6.23-2.0.0

Note: the Buildroot package version name before this commit was
"lf-6.6.3_1.0.0" (with an underscore "_"), which was in fact a
branch git reference. See:
https://github.com/nxp-imx/imx-mkimage/tree/lf-6.6.3_1.0.0

It should have been the tag name "lf-6.6.3-1.0.0" (with the dash "-").
See:
https://github.com/nxp-imx/imx-mkimage/releases/tag/lf-6.6.3-1.0.0

This commit fixes this issue by updating to a new git tag reference.

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>