configs/mx6sx_udoo_neo: bump kernel to 6.6.43
The current 5.15.13 kernel fails to build:

  In file included from <command-line>:
  drivers/ata/libahci.c: In function 'ahci_led_store':
  ././include/linux/compiler_types.h:322:45: error: call to '__compiletime_assert_329' declared with attribute error: BUILD_BUG_ON failed: sizeof(_s) > sizeof(long)

This commit fixes this by using the opportunity to bump the kernel to
the latest LTS version. While bumping the kernel, the headers to use
are also updated, and the selected device trees are updated to use the
new dts per vendor organization.

This update does not introduce any new error or warning with the
current configuration.

Fixes:

  https://gitlab.com/buildroot.org/buildroot/-/jobs/7477108980

Signed-off-by: Philip-Dylan Gleonec <philip-dylan@gleonec.bzh>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>