package/xz: avoid cicular dependency on autoconf for host builds
autoconf has an implicit dependency on HOST_XZ due to extracting a
tar.xz archive. By enabling XZ_AUTORECONF we create a reverse dependency
for xz on autoconf. This is fine for target builds, but it won't work
for host builds, as the two packages now depend on each other.

This problem only occurs if host tools are being built by Buildroot. If
the build relies on host tools provided by the distro, there is no
circular dependency problem.

Fixes:
4703a7ae9176 package/xz: fix musl static build

Signed-off-by: Markus Mayer <mmayer@broadcom.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>