board/raspberrypi: reword console comment
The Raspberry Pi with Bluetooth connectivity up to 4 (i.e. Zero W, Zero
2 W, 3, 4 and Compute Module 4) use the miniuart-bt overlay to restore
UART0 for the serial console (like the Raspberry Pi and Raspberry Pi 2
that have no Bluetooth); the Bluetooth uses the mini-UART instead[1][2].

This rewords the comment instead of enumerating every hardwares that can
be extended in the future.

[1]: https://www.raspberrypi.com/documentation/computers/configuration.html#uarts-and-device-tree
[2]: https://github.com/raspberrypi/linux/blob/rpi-6.6.y/arch/arm/boot/dts/overlays/miniuart-bt-overlay.dts#L4-L6

Signed-off-by: Gaël PORTAY <gael.portay@rtone.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>