package/ffmpeg: fix microblaze compile issue
Since the update of ffmpeg to 6.1.1 in commit
fa36a6375dbeae131c596070570cec66c7bd2456 there are
a lot of microblaze build failures.

The compilation failure is triggered by GCC Bug 68485.

We need to disable optimisations, or ffmpeg would add -O3, which would
be present after the -O0 we add, overriding it.

Fixes:
 - http://autobuild.buildroot.net/results/524/524cef193aec30471e55eb0eecfa39f0d6011a91

And many more.

Signed-off-by: Waldemar Brodkorb <wbx@openadk.org>
[yann.morin.1998@free.fr: explain why --disable-optimizations]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>