package: move wireplumber from Graphics to Audio/Video
WirePlumber is a PipeWire session manager, essential to the operation
of PipeWire in most use cases. It is not a graphical application,
nor does it support graphics in any way. It is only there to support
PipeWire and it should be kept together with PipeWire in the menus.

Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>
(cherry picked from commit 14b1b8ca8dc92851aa7400af432874621cfdb776)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>