package/pipewire: add option to build the AAC codec in spa-bluez
This allows for more precise control of the AAC codec feature,
instead of it being silently enabled if the fdk-aac library happens
to be built earlier, or disabled otherwise.

Also remove the -Dbluez5-codec-opus=disabled option in the case where
the entire bluez5 integration is disabled. -Dbluez5=disabled implies
that all codecs are also disabled.

Signed-off-by: George Kiagiadakis <george.kiagiadakis@collabora.com>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>