package/bcc: bump version to 0.31.0
For a change log since 0.29.1, see:
https://github.com/iovisor/bcc/blob/v0.31.0/debian/changelog

This commit also changes BCC_SITE_METHOD to "git". This is because the
upstream source archive name does not contain the version. It is
always named "bcc-src-with-submodule.tar.gz" in all releases. If using
the https _SITE_METHOD, bumping the version will not download the new
version if the directory BR2_DL_DIR contains an old one, with the same
name. Since the upstream repository uses git submodules, this commit
also adds "BCC_GIT_SUBMODULES = YES".

Signed-off-by: Julien Olivain <ju.o@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>