board/friendlyarm/nanopi-neo: refine genimage.cfg
This patch adds the offset argument for rootfs partiton so that it's
aligned to a 1MiB boundary, instead of using the size of the previous
u-boot partition to achieve this.

Genimage will make sure no partitions overlap. So we don't need a size
argument to limit the size of the u-boot image.

Signed-off-by: Dong Wang <wangdong115@foxmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>