package/fan-ctrl: remove package
For this package, we're downloading from the CVS repository viewer,
but it is gone. The only way would be to download from the CVS
repository itself, but really upstream is completely dead, and this
package has seen no update since it has been introduced in 2013, so
let's get rid of it.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Acked-by: Peter Korsgaard <peter@korsgaard.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>