package/python-django: bump to 5.1
Release notes: https://docs.djangoproject.com/en/5.1/releases/5.1/

We need to add --skip-dependency-check to build options as django
currently pins setuptools <69.3 [1] and buildroot uses a newer version.

The Django pin is likely to not be affected by PEP-625 [2] handling,
which was added to setuptools 69.3 [3][4]. We don't really care about
the sdist name changing for django though, so we can use a newer version
of setuptools as well.

Django has been confirmed to still install and work correctly by running
the runtime test.

[1] https://github.com/django/django/commit/4686541691dbe986f58ac87630c3b7a04db4ff93
[2] https://peps.python.org/pep-0625/
[3] https://github.com/pypa/setuptools/issues/3593
[4] https://github.com/pypa/setuptools/blob/main/NEWS.rst#v6930

Signed-off-by: Marcus Hoffmann <buildroot@bubu1.eu>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>