package/libuecc: switch to a functional Github repo
The https://projects.universe-factory.net/attachments/download/85 URL
no longer works, so let's switch to what appears to be the official
Github repo:

  https://github.com/neocturne/libuecc

The v7 in this Github is 100% identical to the v7 that was available
from https://projects.universe-factory.net/attachments/download/85.

Also at
https://metadata.ftp-master.debian.org/changelogs//main/libu/libuecc/libuecc_7-5_copyright
Debian lists https://github.com/NeoRaider/libuecc as the source, and
this URL redirects to https://github.com/neocturne/libuecc.

Thanks to the original tarball being a .tar.xz, and our Github macro
generating a .tar.gz, the filename is different, and therefore the
different hash will not be an issue.

Fixes:

  http://autobuild.buildroot.org/results/c4c24bd06bc7889c46d78e14d5bef528340d4463/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>