package/gobject-introspection: use objdump instead of prelink-cross
This commit switches the gobject-introspection package from using
prelink-cross to simply using "objdump -p". This basically implements
the same change as was done in OpenEmbedded commit:

  https://git.openembedded.org/openembedded-core/commit/?id=767e0880d4d729e659e859dd99c1cdb084b8ba51

In particular, this allows to get rid of the architecture dependency
of prelink-cross, which wasn't currently encoded, but that was
actually causing failures such as:

/home/autobuild/autobuild/instance-17/output-1/host/sbin/prelink-rtld: error while loading shared libraries: /lib/libc.so.6: cannot handle TLS data

It also paves the way for removing host-prelink-cross entirely, as the
only remaining user of it now is dracut.

The test tests.package.test_python_gobject.TestPythonPy3Gobject (which
uses gobject-introspection) has been tested to run fine after this
commit.

Fixes:

  http://autobuild.buildroot.net/results/7efa6a18f543752dfe33577f74c20bda3f8b1e65/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Tested-by: Fiona Klute (WIWA) <fiona.klute@gmx.de>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>